requirements.txt
python-telegram-bot>=22.0
feedparser>=6.0
twikit>=2.0
google-cloud-firestore>=2.15.0
python-dotenv>=1.0
starlette>=0.35.0
uvicorn>=0.27.0
deep-translator>=1.11.0
apscheduler=3.11.2
